% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_extract_csv_tables.R
\name{add_keys}
\alias{add_keys}
\title{Add KEYS columns to a flattened table}
\usage{
add_keys(db_tbl, table_name, year, cc_file, con)
}
\arguments{
\item{db_tbl}{Lazy tibble (or data frame).}

\item{table_name}{Character table id.}

\item{year}{Integer year.}

\item{cc_file}{Concordance crosswalk data frame.}

\item{con}{DBI connection.}
}
\value{
A tibble with KEYS columns relocated first.
}
\description{
Add KEYS columns to a flattened table
}
