% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_write_to_duckdb.R
\name{check_for_columns}
\alias{check_for_columns}
\title{Ensure destination DuckDB table has all required columns}
\usage{
check_for_columns(new_data, table_name, con)
}
\arguments{
\item{new_data}{Data frame with new data to append.}

\item{table_name}{Character table name.}

\item{con}{DBI connection to DuckDB.}
}
\description{
Adds any missing columns (as TEXT) needed to append \code{new_data} into an
existing DuckDB table.
}
