% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_release_notes.R
\name{condense_years}
\alias{condense_years}
\title{Collapse years into ranges}
\usage{
condense_years(y)
}
\arguments{
\item{y}{Integer years.}
}
\value{
A string such as \code{"2009-2011, 2013"}.
}
\description{
Collapse years into ranges
}
\examples{
condense_years( c( 2009, 2010, 2011, 2013 ) )
}
