% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_s3.R
\name{configure_aws_credentials}
\alias{configure_aws_credentials}
\title{Configure AWS credentials for DuckDB session}
\usage{
configure_aws_credentials(con)
}
\arguments{
\item{con}{DBI connection to DuckDB.}
}
\description{
Configure AWS credentials for DuckDB session
}
