% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_batchfiles.R
\name{create_batchfiles}
\alias{create_batchfiles}
\title{Create batchfiles (RDS) for multiple years}
\usage{
create_batchfiles(index, years, group.size)
}
\arguments{
\item{index}{Data frame with URL and TaxYear.}

\item{years}{Vector of years (integer or character).}

\item{group.size}{Integer batch size.}
}
\description{
Create batchfiles (RDS) for multiple years
}
