% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_write_table_output.R
\name{dbq}
\alias{dbq}
\title{Quote a SQL identifier}
\usage{
dbq(x)
}
\arguments{
\item{x}{Character vector of identifiers.}
}
\value{
Character vector of double-quoted identifiers, embedded quotes doubled.
}
\description{
Quote a SQL identifier
}
\keyword{internal}
