% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{download_current_index_full}
\alias{download_current_index_full}
\title{Download Current AWS Index}
\usage{
download_current_index_full()
}
\value{
None. Downloads the file as "INDEX.CSV" in the working directory.
}
\description{
Downloads the most recent AWS index file.
}
\examples{
\dontrun{
fn <- download_current_index_full()
index <- data.table::fread( fn )
split_index( index )
}
}
