% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef2-package.R
\docType{package}
\name{ef2-package}
\alias{ef2}
\alias{ef2-package}
\title{ef2: Tools for Building and Maintaining the IRS 990 Efiler Database}
\description{
Provides tools for retrieving, parsing, and flattening IRS 990 e-file XML data into relational form, including schema alignment, batch processing, and DuckDB integration. Supports large-scale ingestion and alignment of IRS e-file data using concordance crosswalks and parallelized retrieval.
}
\author{
\strong{Maintainer}: Jesse Lecy \email{jlecy@asu.edu}

}
\keyword{internal}
