% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{extract_dates}
\alias{extract_dates}
\title{Extract Dates from Filenames}
\usage{
extract_dates(x)
}
\arguments{
\item{x}{A character vector containing filenames.}
}
\value{
A character vector of extracted dates.
}
\description{
Extracts dates in "YYYY-MM-DD" format from a vector of strings.
}
\examples{
extract_dates("xmls_created_on_2023-11-19.csv")
}
