% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{extract_filenames_batch}
\alias{extract_filenames_batch}
\title{Extract Batch Index Filenames}
\usage{
extract_filenames_batch(strings)
}
\arguments{
\item{strings}{A character vector containing strings to search.}
}
\value{
A character vector of matched filenames.
}
\description{
Extracts batch index filenames from the list of all files in the GTDC AWS S3 index bucket.
}
\examples{
\dontrun{
index.list <- get_index_list_awscli()
extract_filenames_batch( index.list )
}
}
