% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{extract_filenames_full}
\alias{extract_filenames_full}
\title{Extract Index Filenames}
\usage{
extract_filenames_full(strings)
}
\arguments{
\item{strings}{A character vector containing strings to search.}
}
\value{
A character vector of matched filenames.
}
\description{
Extracts filenames matching a specific pattern from a vector of strings.
}
\examples{
\dontrun{
index.list <- get_index_list_awscli()
extract_filenames_full( index.list )
}
}
