% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{find_parent_nodes}
\alias{find_parent_nodes}
\title{Find parent node xpaths for a set of xpaths}
\usage{
find_parent_nodes(xpath_list)
}
\arguments{
\item{xpath_list}{Character vector.}
}
\value{
Character vector of unique parent xpaths.
}
\description{
Find parent node xpaths for a set of xpaths
}
