% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{find_terminal_nodes}
\alias{find_terminal_nodes}
\title{Find terminal node xpaths for a set of xpaths}
\usage{
find_terminal_nodes(xpath_list)
}
\arguments{
\item{xpath_list}{Character vector.}
}
\value{
Character vector of terminal xpaths.
}
\description{
Find terminal node xpaths for a set of xpaths
}
