% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_extract_csv_tables.R
\name{flatten_table}
\alias{flatten_table}
\title{Flatten a logical RDB table into wide format from FLATXML}
\usage{
flatten_table(table_name, year, con)
}
\arguments{
\item{table_name}{Character table identifier in the concordance.}

\item{year}{Integer year.}

\item{con}{DBI connection to DuckDB.}
}
\value{
A lazy tibble (dbplyr) that can be \code{collect()}ed.
}
\description{
Flatten a logical RDB table into wide format from FLATXML
}
