% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concordance.R
\name{get_concordance}
\alias{get_concordance}
\title{Retrieve the Concordance File}
\usage{
get_concordance(gh = TRUE)
}
\arguments{
\item{gh}{Logical; if TRUE, fetches the concordance from GitHub.}
}
\value{
A data.table or data.frame containing the concordance.
}
\description{
Downloads the master concordance file from GitHub or loads a packaged version.
}
