% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_header}
\alias{get_header}
\title{Compute the TABLE_HEADER from an xpath}
\usage{
get_header(xpath, type)
}
\arguments{
\item{xpath}{Character scalar xpath.}

\item{type}{One of "parent" or "terminal".}
}
\value{
Character scalar header xpath (two-level context).
}
\description{
Compute the TABLE_HEADER from an xpath
}
