% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{get_index_list_awscli}
\alias{get_index_list_awscli}
\title{Get AWS Index List}
\usage{
get_index_list_awscli()
}
\value{
A character vector containing the raw output of the AWS CLI command.
}
\description{
Retrieves a list of AWS indices using the AWS CLI.
}
\examples{
\dontrun{
get_index_list_awscli()
}
}
