% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{get_last_n_dates}
\alias{get_last_n_dates}
\title{Get Last N Dates}
\usage{
get_last_n_dates(N = 30)
}
\arguments{
\item{N}{An integer specifying the number of days to generate.}
}
\value{
A character vector of dates in "YYYY-MM-DD" format.
}
\description{
Generates a sequence of the last N dates in "YYYY-MM-DD" format.
}
\examples{
get_last_n_dates(30)
}
