% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_n}
\alias{get_n}
\title{Extract the last bracketed index from an xpath}
\usage{
get_n(x)
}
\arguments{
\item{x}{Character scalar xpath.}
}
\value{
Character index (defaults to "0" if none).
}
\description{
For an xpath like "\verb{/Return/.../ScheduleO[3]/.../Line[12]}" returns "12".
}
