% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_diagram_table_structure.R
\name{get_table_xpaths}
\alias{get_table_xpaths}
\title{Get Table XPaths}
\usage{
get_table_xpaths(table.name = "F9-P07-T01-COMPENSATION")
}
\arguments{
\item{table.name}{A character string specifying the table name.}
}
\value{
A character vector of XPaths associated with the table name.
}
\description{
Retrieves the XPaths associated with a specific table name from the concordance.
}
\examples{
get_table_xpaths("F9-P03-T01-PROGRAMS-OTHER")
}
