% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_type}
\alias{get_type}
\title{Classify xpaths as parent or terminal}
\usage{
get_type(xpath)
}
\arguments{
\item{xpath}{Character vector of xpaths.}
}
\value{
Character vector with values "parent" or "terminal".
}
\description{
Classify xpaths as parent or terminal
}
