% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{get_url_status}
\alias{get_url_status}
\title{Get URL Status}
\usage{
get_url_status(url)
}
\arguments{
\item{url}{A character string representing the URL to check.}
}
\value{
A data frame containing the URL, its existence status, and the HTTP status code.
}
\description{
Retrieves the HTTP status of a given URL.
}
\examples{
get_url_status("https://example.com")
}
