% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_vnames}
\alias{get_vnames}
\title{Vectorized variable name extraction from xpaths}
\usage{
get_vnames(xpath, type)
}
\arguments{
\item{xpath}{Character vector.}

\item{type}{Character vector (ignored; reserved for future behavior).}
}
\value{
Character vector of variable names.
}
\description{
Vectorized variable name extraction from xpaths
}
