% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_xpath_vname}
\alias{get_xpath_vname}
\title{Get the last node name from an xpath}
\usage{
get_xpath_vname(x)
}
\arguments{
\item{x}{Character scalar xpath.}
}
\value{
Character node name.
}
\description{
Get the last node name from an xpath
}
