% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_patch_index.R
\name{irs_xml_base}
\alias{irs_xml_base}
\title{Base URL of the IRS e-file XML downloads}
\usage{
irs_xml_base()
}
\value{
Character URL with a trailing slash.
}
\description{
Base URL of the IRS e-file XML downloads
}
\keyword{internal}
