% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concordance.R
\name{prep_concordance}
\alias{prep_concordance}
\title{Prepare a concordance crosswalk (uppercase colnames)}
\usage{
prep_concordance(ccf = NULL)
}
\arguments{
\item{ccf}{Optional concordance; if NULL, loads via \code{get_concordance()}.}
}
\value{
Data frame with columns XPATH, VARIABLE_NAME, RDB_TABLE.
}
\description{
Prepare a concordance crosswalk (uppercase colnames)
}
