% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_release_notes.R
\name{release_files}
\alias{release_files}
\title{List the Parquet tables of a release}
\usage{
release_files(path)
}
\arguments{
\item{path}{A local folder, an \verb{s3://bucket/prefix/} URI, or a bare prefix
of the NCCS bucket such as \code{"public/efile_v3_1/"}. S3 prefixes are listed
anonymously over HTTPS.}
}
\value{
A \code{data.table} with \code{table}, \code{year} and \code{file} (a path or HTTPS URL
DuckDB can read).
}
\description{
Lists \verb{<TABLE>-<YEAR>.parquet} files in a local folder or an
S3 prefix of the public NCCS bucket.
}
\keyword{internal}
