% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_write_to_duckdb.R
\name{send_flat_xml_to_db}
\alias{send_flat_xml_to_db}
\title{Write flattened XML batch results to an existing DuckDB connection}
\usage{
send_flat_xml_to_db(RESULTS, con)
}
\arguments{
\item{RESULTS}{List of parsed XML outputs from \code{get_flat_xml()}.}

\item{con}{Active DBI connection to DuckDB (persistent).}
}
\value{
Invisibly TRUE on success.
}
\description{
Write flattened XML batch results to an existing DuckDB connection
}
