% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_batchfiles.R
\name{split_index}
\alias{split_index}
\title{Split index to batchfile RDS for a single year}
\usage{
split_index(year, index, group.size = 25)
}
\arguments{
\item{year}{Integer/character year.}

\item{index}{Data frame.}

\item{group.size}{Integer batch size.}
}
\value{
Invisibly the batch list.
}
\description{
Split index to batchfile RDS for a single year
}
