% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_irs990efile.R
\name{standardize_boole}
\alias{standardize_boole}
\title{Standardize boolean inputs}
\usage{
standardize_boole(x)
}
\arguments{
\item{x}{A vector of values to standardize.}
}
\value{
A logical vector.
}
\description{
Converts various string representations of true/false (e.g. "YES", "X", "1")
into a logical vector. \code{NA} values are treated as \code{FALSE}.
}
\examples{
standardize_boole(c("YES", "NO", NA, "X"))
}
