% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_extract_csv_tables.R
\name{build_rdb_table}
\alias{build_rdb_table}
\title{Build an RDB table from multiple header variants}
\usage{
build_rdb_table(
  table_name,
  year,
  TABLE.HEADERS,
  con,
  cc_file,
  post_to_s3 = FALSE
)
}
\arguments{
\item{table_name}{Character table id.}

\item{year}{Integer year.}

\item{TABLE.HEADERS}{Named list of header xpaths per table.}

\item{con}{DBI connection.}

\item{cc_file}{Concordance crosswalk.}

\item{post_to_s3}{Logical export flag.}
}
\value{
Invisibly the lazy tibble.
}
\description{
Build an RDB table from multiple header variants
}
