% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_extract_csv_tables.R
\name{build_table}
\alias{build_table}
\title{Build a structured wide table and optionally export to CSV/S3}
\usage{
build_table(table_name, year, con, cc_file, post_to_s3 = FALSE)
}
\arguments{
\item{table_name}{Character table id.}

\item{year}{Integer year.}

\item{con}{DBI connection.}

\item{cc_file}{Concordance crosswalk.}

\item{post_to_s3}{Logical, if TRUE write CSV to S3 using DuckDB COPY.}
}
\value{
Invisibly the lazy tibble.
}
\description{
Build a structured wide table and optionally export to CSV/S3
}
