% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_get_gt_index.R
\name{find_current_index_full}
\alias{find_current_index_full}
\title{Find Most Recent AWS Full Index}
\usage{
find_current_index_full(days = 100)
}
\arguments{
\item{days}{An integer specifying the number of past days to check.}
}
\value{
A character string representing the URL of the most recent index file, or NA if none found.
}
\description{
Identifies the most recent AWS index file (all years) within the given number of days.
}
\examples{
find_current_index_full(100)
}
