% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_irs990efile.R
\name{get_object_id}
\alias{get_object_id}
\title{Extract an OBJECTID from a filing URL}
\usage{
get_object_id(url)
}
\arguments{
\item{url}{Character. Full XML URL.}
}
\value{
Character scalar OBJECTID (e.g. \code{"OID-202301529349200315"}).
}
\description{
Parses the unique object ID from an IRS 990 e-file XML URL and prefixes it
with \verb{OID-}.
}
\examples{
base <- "https://gt990datalake-rawdata.s3.amazonaws.com/EfileData/XmlFiles/"
get_object_id(paste0(base, "202301529349200315_public.xml"))
# "OID-202301529349200315"
}
\seealso{
\code{\link[=get_object_id2]{get_object_id2()}} for a variant that also handles the TEOS xml2 URL bases.
}
