% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_object_id2}
\alias{get_object_id2}
\title{Extract an OBJECTID from a filing URL}
\usage{
get_object_id2(url)
}
\arguments{
\item{url}{Character. Full XML URL.}
}
\value{
Character scalar OBJECTID.
}
\description{
Converts known IRS 990 e-file XML URLs into a standardized OBJECTID
(prefixed with \verb{OID-}) used as a database key.
}
\examples{
get_object_id2("https://nccs-efile.s3.us-east-1.amazonaws.com/xml/202220139349301207_public.xml")
}
