% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_s3.R
\name{get_s3_database}
\alias{get_s3_database}
\title{Attach an S3-hosted DuckDB database by filename}
\usage{
get_s3_database(filename, anonymous = TRUE)
}
\arguments{
\item{filename}{DuckDB filename within s3://nccs-efile/duckdb/.}

\item{anonymous}{Logical for anonymous access.}
}
\value{
DBI connection with attached database.
}
\description{
Attach an S3-hosted DuckDB database by filename
}
