% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_table_id}
\alias{get_table_id}
\title{Make a TABLE_ID from a vector of xpaths}
\usage{
get_table_id(xpaths)
}
\arguments{
\item{xpaths}{Character vector of xpaths.}
}
\value{
Character vector like "TID-00003".
}
\description{
Make a TABLE_ID from a vector of xpaths
}
