% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils_s3.R
\name{open_database}
\alias{open_database}
\title{Open a DuckDB database connection with S3 support}
\usage{
open_database(s3_region = "us-east-1", anonymous = TRUE)
}
\arguments{
\item{s3_region}{AWS region (default "us-east-1").}

\item{anonymous}{Logical; TRUE for anonymous S3 access.}
}
\value{
DBI connection.
}
\description{
Open a DuckDB database connection with S3 support
}
