% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_batchfiles.R
\name{split_into_groups}
\alias{split_into_groups}
\title{Utility: split a vector into labeled groups like "G01{25}"}
\usage{
split_into_groups(x, G = 25)
}
\arguments{
\item{x}{Vector to split.}

\item{G}{Group size.}
}
\value{
Named list of groups.
}
\description{
Utility: split a vector into labeled groups like "G01{25}"
}
