% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_update_db.R
\name{update_db}
\alias{update_db}
\title{Update the DuckDB database for a given tax year}
\usage{
update_db(year, index, path = ".")
}
\arguments{
\item{year}{Integer tax year.}

\item{index}{Data frame with TaxYear and URL columns.}
}
\value{
Invisibly path to merged database or NULL.
}
\description{
Update the DuckDB database for a given tax year
}
